% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AddAnnotation}
\alias{AddAnnotation}
\title{AddAnnotation}
\usage{
AddAnnotation(
  p,
  fsom,
  layout = fsom$MST$l,
  cl = NULL,
  mcl = NULL,
  clCustomLabels = NULL,
  mclCustomLabels = NULL,
  hjust = 0.5,
  ...
)
}
\arguments{
\item{p}{Plot to add annotation to. When using type = "stars", please 
use plot = FALSE in \code{\link{PlotFlowSOM}}.}

\item{fsom}{FlowSOM object that goes with the plot}

\item{layout}{Layout to use. Default fsom$MST$l}

\item{cl}{Clusters to annotate. If cl = "all", all clusters will be 
annotated}

\item{mcl}{Metaclusters to annotate. If mcl = "all", all metaclusters 
will be annotated}

\item{clCustomLabels}{Cluster labels to use for annotation}

\item{mclCustomLabels}{Metacluster labels to use for annotation}

\item{hjust}{Label adjustment. Default = 0.5}

\item{...}{Arguments passed to geom_text_repel}
}
\value{
The updated plot
}
\description{
Add annotation to a FlowSOM plot
}
