% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AddBackground}
\alias{AddBackground}
\title{AddBackground}
\usage{
AddBackground(
  p,
  backgroundValues,
  backgroundColors = NULL,
  backgroundLim = NULL
)
}
\arguments{
\item{p}{GGplot object}

\item{backgroundValues}{Vector of values to be plotted as background for 
the nodes}

\item{backgroundColors}{Colorpalette for backgroundcolors}

\item{backgroundLim}{Background limits (can be used to ensure consistent
colorpalette between plots). If NULL (default), will
be automatically adapted to the data.}
}
\value{
Returns nothing, but plots the background
}
\description{
Function plots the background
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{AddLabels}},
\code{\link{AddNodes}}, \code{\link{AddPies}}, \code{\link{AddStars}}
}
