% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_readInput.R
\name{AddFlowFrame}
\alias{AddFlowFrame}
\title{Add a flowFrame to the data variable of the FlowSOM object}
\usage{
AddFlowFrame(fsom, flowFrame)
}
\arguments{
\item{fsom}{FlowSOM object, as constructed by the ReadInput function}

\item{flowFrame}{flowFrame to add to the FlowSOM object}
}
\value{
FlowSOM object with data added
}
\description{
Add a flowFrame to the data variable of the FlowSOM object
}
\seealso{
\code{\link{ReadInput}}
}
