% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AddLabels}
\alias{AddLabels}
\title{AddLabels}
\usage{
AddLabels(
  p,
  labels,
  hjust = 0.5,
  layout = NULL,
  textSize = 3.88,
  textColor = "black",
  ...
)
}
\arguments{
\item{p}{ggplot object}

\item{labels}{Labels to be added to each node}

\item{hjust}{Horizontal adjust for labels. Default is centered.}

\item{layout}{Dataframe with x and y columns. If null, the dataframe from
the ggplot object will be reused.}

\item{textSize}{Size for geom_text. Default (=3.88) is from geom_text.}

\item{textColor}{Color for geom_text. Default = black.}

\item{...}{Additional parameters to pass to geom_text}
}
\value{
Returns the ggplot object with labels added
}
\description{
AddLabels
}
\seealso{
\code{\link{PlotLabels}}, \code{\link{PlotNumbers}}
}
