% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AddMST}
\alias{AddMST}
\title{AddMST}
\usage{
AddMST(p, fsom)
}
\arguments{
\item{p}{ggplot object}

\item{fsom}{FlowSOM object, as generated by \code{\link{FlowSOM}}}
}
\value{
Returns nothing, but plots the MST for FlowSOM MST view
}
\description{
Function plots the MST
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{ParseEdges}}, 
\code{\link{AddStarsPies}}, \code{\link{AddLabels}}, \code{\link{AddNodes}},
\code{\link{AddBackground}}, \code{\link{AddPies}}, \code{\link{AddStars}}
}
