% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AddNodes}
\alias{AddNodes}
\title{AddNodes}
\usage{
AddNodes(
  p,
  nodeInfo = NULL,
  values = NULL,
  lim = NULL,
  colorPalette = NULL,
  fillColor = "white",
  showLegend = TRUE,
  label = "",
  ...
)
}
\arguments{
\item{p}{GGplot object}

\item{nodeInfo}{Dataframe with for every node an x, y and size value,
if null the dataframe from the ggplot object will be 
reused.}

\item{values}{Values used for coloring the nodes. Default = NULL,
in which case all nodes are filled in fillColor.}

\item{lim}{The limits of the color scale, not used if values = NULL.}

\item{colorPalette}{Colorpalette for color in nodes, not used if values = 
NULL. A vector of colors or a color function.}

\item{fillColor}{Fixed fill for node colors, default = white.}

\item{showLegend}{Boolean, default = TRUE.}

\item{label}{Title for the legend.}

\item{...}{Additional arguments to pass to geom_circle}
}
\value{
Returns nothing, but plots the nodes
}
\description{
Function plots the nodes
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{PlotMarker}},
\code{\link{PlotVariable}}, \code{\link{AddLabels}}, 
\code{\link{AddBackground}}, \code{\link{AddPies}}, 
\code{\link{AddStars}}, \code{\link{AddStarsPies}}
}
