% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AddPies}
\alias{AddPies}
\title{AddPies}
\usage{
AddPies(p, fsom, cellLabels, layout = NULL, colorPalette = NULL)
}
\arguments{
\item{p}{GGplot object}

\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{cellLabels}{Array of factors indicating the cell labels}

\item{layout}{Coordinates of nodes. Uses dataframe of the ggplot 
object if NULL.}

\item{colorPalette}{ColorPalette to be used for colors. Can be either a 
function or an array specifying colors.}
}
\value{
ggplot object with the pies added
}
\description{
Function plots the pies
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{AddLabels}},
\code{\link{AddNodes}}, \code{\link{AddBackground}}, \code{\link{PlotPies}},
\code{\link{AddStars}}, \code{\link{ParseArcs}}
}
