% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AddScale}
\alias{AddScale}
\title{AddScale}
\usage{
AddScale(
  p,
  values = NULL,
  colors = NULL,
  limits = NULL,
  showLegend = TRUE,
  labelLegend = "",
  type = "fill"
)
}
\arguments{
\item{p}{ggplot object}

\item{values}{Values used for the fill}

\item{colors}{Colors to use (can be a vector or a function)}

\item{limits}{Limits to use in the scale}

\item{showLegend}{Boolean on whether to show the legend}

\item{labelLegend}{Label to show as title of the legend}

\item{type}{fill (default) or color}
}
\value{
ggplot object with scale added
}
\description{
AddScale
}
