% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AddStars}
\alias{AddStars}
\title{AddStars}
\usage{
AddStars(p, fsom, markers = fsom$map$colsUsed, colorPalette = NULL)
}
\arguments{
\item{p}{GGplot object}

\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{markers}{Determines which markers to plot. 
Default = "fsom$map$colsUsed"}

\item{colorPalette}{ColorPalette to be used for colors. Can be either a 
function or an array specifying colors.}
}
\value{
ggplot object with the stars added
}
\description{
Function plots the stars
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{AddLabels}},
\code{\link{AddNodes}}, \code{\link{AddBackground}}, \code{\link{PlotStars}},
\code{\link{AddPies}}, \code{\link{ParseArcs}}
}
