% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AddStarsPies}
\alias{AddStarsPies}
\title{AddStarsPies}
\usage{
AddStarsPies(p, arcs, colorPalette, showLegend = TRUE)
}
\arguments{
\item{p}{ggplot object}

\item{arcs}{Dataframe that contains all the data for the plotting 
the pies or stars}

\item{colorPalette}{A vector of colors or a color function}

\item{showLegend}{Boolean on whether to show the legend}
}
\value{
Returns nothing, but plots the stars or pies
}
\description{
Function plots stars or pies
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{AddLabels}},
\code{\link{AddNodes}}, \code{\link{AddBackground}}, \code{\link{AddPies}},
\code{\link{AddStars}}, \code{\link{ParseArcs}}, \code{\link{PlotStars}}
\code{\link{PlotPies}}
}
