% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{AutoMaxNodeSize}
\alias{AutoMaxNodeSize}
\title{AutoMaxNodeSize}
\usage{
AutoMaxNodeSize(layout, overlap)
}
\arguments{
\item{layout}{Coordinates of nodes}

\item{overlap}{Parameter that determines how much overlap there 
will be. If negative the nodes will be smaller}
}
\value{
Returns the maxNodeSize with some overlap
}
\description{
Calculate node size
}
\details{
Function that calculates the minimum distance between the nodes
to use this to adapt the maxNodeSize for better plotting
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{ScaleStarHeights}}, 
\code{\link{ParseNodeSize}}
}
