\name{BuildMST}
\alias{BuildMST}
\title{Build Minimal Spanning Tree}
\description{Add minimal spanning tree description to the FlowSOM object}
\usage{BuildMST(fsom, silent=FALSE)}
\arguments{
    \item{fsom}{FlowSOM object, as generated by \code{\link{BuildSOM}}}
    \item{silent}{if \code{TRUE}, no progress updates will be printed}
}
\value{FlowSOM object containing MST description}
\seealso{
    \code{\link{BuildSOM}}
}
\examples{
    # Read from file, build self-organizing map
    fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
    flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                            scale=TRUE)
    flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
    
    # Build the Minimal Spanning Tree
    flowSOM.res <- BuildMST(flowSOM.res)
}