% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{computeBackgroundColor}
\alias{computeBackgroundColor}
\title{Internal function for computing background nodes}
\usage{
computeBackgroundColor(
  backgroundValues,
  backgroundColor,
  backgroundLim = NULL,
  backgroundBreaks = NULL
)
}
\arguments{
\item{backgroundValues}{Values to be used for background coloring, either
numerical values or something that can be made into
a factor (e.g. a clustering)}

\item{backgroundColor}{Colorpalette to be used for the background coloring
. Can be either a function or an array specifying
colors}

\item{backgroundLim}{Only used when backgroundValues are numerical. 
Defaults to min and max of the backgroundValues.}

\item{backgroundBreaks}{Breaks to pass on to \code{\link{cut}}, to split
numerical background values. If NULL, the length of
backgroundColor will be used (default 100).}
}
\description{
Internal function for computing background nodes
}
