% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2_buildSOM.R
\name{CountGroups}
\alias{CountGroups}
\title{Calculate differences in cell counts between groups}
\usage{
CountGroups(fsom, groups, plot = TRUE, silent = FALSE)
}
\arguments{
\item{fsom}{FlowSOM object as generated by BuildSOM}

\item{groups}{List containing an array with file names for each group}

\item{plot}{Logical. If TRUE, make a starplot of each individual file}

\item{silent}{Logical. If TRUE, print progress messages}
}
\value{
Distance matrix
}
\description{
Calculate differences in cell counts between groups
}

