% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_buildSOM.R
\name{Dist.MST}
\alias{Dist.MST}
\title{Calculate distance matrix using a minimal spanning tree neighbourhood}
\usage{
Dist.MST(X)
}
\arguments{
\item{X}{matrix in which each row represents a point}
}
\value{
Distance matrix
}
\description{
Calculate distance matrix using a minimal spanning tree neighbourhood
}
