% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{FlowSOM_colors}
\alias{FlowSOM_colors}
\title{FlowSOM default colors}
\usage{
FlowSOM_colors(n)
}
\arguments{
\item{n}{Number of colors to generate}
}
\value{
array of n colors
}
\description{
FlowSOM default colors
}
