% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_deprecated.R
\name{get_channels}
\alias{get_channels}
\title{get_channels}
\usage{
get_channels(ff, markers)
}
\arguments{
\item{ff}{The flowFrame of interest}

\item{markers}{Vector with markers or channels of interest}
}
\value{
Corresponding channel names
}
\description{
Get channel names for an array of markers, given a flowframe
}
\examples{

   # Read the flowFrame
   fileName <- system.file("extdata", "68983.fcs", package="FlowSOM")
   ff <- flowCore::read.FCS(fileName)
   GetChannels(ff, c("FSC-A", "CD3", "FITC-A"))
   GetMarkers(ff, c("FSC-A", "CD3", "FITC-A"))

}
\seealso{
\code{\link{get_markers}}
}
