% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_FlowSOM.R
\name{get_markers}
\alias{get_markers}
\title{get_markers}
\usage{
get_markers(ff, markers)
}
\arguments{
\item{markers}{Vector with markers or channels of interest}

\item{flowFrame}{The flowFrame of interest}
}
\value{
Corresponding marker names
}
\description{
Get marker names, given a flowframe. As available in "desc". If this is NA,
defaults to channel name.
}
\examples{

   # Read the flowFrame
   fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
   ff <- flowCore::read.FCS(fileName)
   get_channels(ff, c("FSC-A", "CD3", "FITC-A"))
   get_markers(ff, c("FSC-A", "CD3", "FITC-A"))

}
\seealso{
\code{\link{get_channels}}
}
