% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_buildSOM.R
\name{GetCVs}
\alias{GetCVs}
\title{Get CV values for all clusters}
\usage{
GetCVs(fsom)
}
\arguments{
\item{fsom}{FlowSOM object as generated by the FlowSOM function
or the BuildSOM function}
}
\value{
Matrix with coefficient of variation values for each marker
}
\description{
Get CV values for all clusters
}
