% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_metaClustering.R
\name{GetMajorityLabels}
\alias{GetMajorityLabels}
\title{GetMajorityLabels}
\usage{
GetMajorityLabels(fsom, labels)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{FlowSOM}}}

\item{labels}{Array of celltypes per cell in \code{fsom}}
}
\value{
Named list with array of majority labels for the FlowSOM clusters and 
metaclusters.
}
\description{
Get majority cluster and metacluster labels from a vector of manual labels 
per cell in the FlowSOM object. Especially useful as starting point for 
\code{\link{UpdateMetaclusters}}.
}
\examples{
# Identify the files
fcs_file <- system.file("extdata", "68983.fcs", package = "FlowSOM")
gating_file <- system.file("extdata", "gatingResult.csv", package = "FlowSOM")

# Specify the cell types of interest for assigning one label per cell
cellTypes <- c("B cells",
               "gd T cells", "CD4 T cells", "CD8 T cells",
               "NK cells", "NK T cells")
                
# Load manual labels (e.g. GetFlowJoLabels can be used to extract labels from
# an fcs file)
gatingResult <- as.factor(read.csv(gating_file, header = FALSE)[, 1])


# Build a FlowSOM tree
flowSOM.res <- FlowSOM(fcs_file,
                       scale = TRUE, 
                       compensate = TRUE, 
                       transform = TRUE,
                       toTransform = 8:18, 
                       colsToUse = c(9, 12, 14:18),
                       nClus = 10,
                       seed = 1)
   
# Get metacluster label based on the gatingResult
majorityLabels <- GetMajorityLabels(fsom = flowSOM.res, 
                                    labels = gatingResult)


}
\seealso{
\code{\link{UpdateMetaclusters}}
}
