% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of colors}
}
\value{
array with hexadecimal color values
}
\description{
Helper function to get the ggplot colors
}
