% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2_buildSOM.R
\name{Initialize}
\alias{Initialize}
\title{Select k well spread points from X}
\usage{
Initialize(X, k)
}
\arguments{
\item{X}{matrix in which each row represents a point}

\item{k}{number of points to choose}
}
\value{
array containing indices of selected rows
}
\description{
Select k well spread points from X
}

