% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_buildSOM.R
\name{Initialize_KWSP}
\alias{Initialize_KWSP}
\title{Select k well spread points from X}
\usage{
Initialize_KWSP(X, xdim, ydim)
}
\arguments{
\item{X}{matrix in which each row represents a point}

\item{xdim}{x dimension of the grid}

\item{ydim}{y dimension of the grid}
}
\value{
array containing the selected selected rows
}
\description{
Select k well spread points from X
}
\examples{

points <- matrix(1:1000, ncol = 10)
selection <- Initialize_KWSP(points, 3, 3)

}
