% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{NClusters}
\alias{NClusters}
\title{NClusters}
\usage{
NClusters(fsom)
}
\arguments{
\item{fsom}{FlowSOM object}
}
\value{
The number of clusters
}
\description{
Extracts the number of clusters from a FlowSOM object
}
