% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{NMetaclusters}
\alias{NMetaclusters}
\title{NMetaclusters}
\usage{
NMetaclusters(fsom)
}
\arguments{
\item{fsom}{FlowSOM object}
}
\value{
The number of metaclusters
}
\description{
Extracts the number of metaclusters from a FlowSOM object
}
