% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{ParseArcs}
\alias{ParseArcs}
\title{ParseArcs}
\usage{
ParseArcs(x, y, arcValues, arcHeights)
}
\arguments{
\item{x}{x coordinate of node}

\item{y}{y coordinate of node}

\item{arcValues}{A named vector with the frequency of how the node 
should be divided}

\item{arcHeights}{The heights of the arcs}
}
\value{
A dataframe ready to use with ggplot, consisting of the coordinates 
        of centers, the radius and angles of the star values
}
\description{
Parses stars
}
\details{
Function that parses the FlowSOM object into a dataframe for the star values 
for ggplot
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{ParseEdges}}, 
\code{\link{ParseNodeSize}},
\code{\link{ParseQuery}}, \code{\link{ParseSD}}
}
