% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{ParseEdges}
\alias{ParseEdges}
\title{ParseEdges}
\usage{
ParseEdges(fsom)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{FlowSOM}}}
}
\value{
A dataframe consisting of start and end coordinates of edges
}
\description{
Parses edges
}
\details{
Function that parses the graph edges of the FlowSOM object into a dataframe
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{ParseNodeSize}},
\code{\link{ParseArcs}}, \code{\link{ParseQuery}},
\code{\link{ParseSD}}, \code{\link{AddMST}}
}
