% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{ParseLayout}
\alias{ParseLayout}
\title{ParseLayout}
\usage{
ParseLayout(fsom, layout)
}
\arguments{
\item{fsom}{FlowSOM object}

\item{layout}{"MST", "grid" or a matrix/dataframe with 2 columns and 1 row
per cluster}
}
\value{
dataframe with 2 columns and 1 row per cluster
}
\description{
ParseLayout
}
