% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{ParseNodeSize}
\alias{ParseNodeSize}
\title{ParseNodeSize}
\usage{
ParseNodeSize(nodeSizes, maxNodeSize, refNodeSize)
}
\arguments{
\item{nodeSizes}{A vector with node sizes}

\item{maxNodeSize}{Determines the maximum node size.}

\item{refNodeSize}{Reference for node size against which the nodeSizes 
will be scaled. Default = max(nodeSizes)}
}
\value{
A vector is returned consisting of node sizes
}
\description{
Parses node size
}
\details{
Function that parses the mapping of the FlowSOM object into node sizes 
relative to the abundances of cells per cluster

Scales node size relative to the abundances of cells per cluster
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{ParseEdges}}, 
\code{\link{AutoMaxNodeSize}}, \code{\link{ParseArcs}},
\code{\link{ParseQuery}}, \code{\link{ParseSD}}
}
