% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{ParseQuery}
\alias{ParseQuery}
\title{ParseQuery}
\usage{
ParseQuery(fsom, query)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{FlowSOM}}}

\item{query}{Array containing "high" or "low" for the specified 
column names of the FlowSOM data}
}
\value{
A list, containing the ids of the selected nodes, the individual  
        scores for all nodes and the scores for each marker for each node
}
\description{
Parses query
}
\details{
Identify nodes in the tree which resemble a certain profile of "high" 
or "low" marker expressions.
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{ParseEdges}}, 
\code{\link{ParseNodeSize}}, \code{\link{ParseArcs}},
\code{\link{QueryStarPlot}}, \code{\link{ParseSD}}
}
