% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{ParseSD}
\alias{ParseSD}
\title{ParseSD 
 
Parses SD in FlowSOM object}
\usage{
ParseSD(fsom, marker = NULL)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{FlowSOM}}}

\item{marker}{If a marker is given, the standard deviation for this marker 
is shown. Otherwise, the maximum ratio is used.}
}
\value{
A vector containing the SDs
}
\description{
Calculates the  standard deviation of a FlowSOM object
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{ParseEdges}}, 
\code{\link{ParseNodeSize}}, \code{\link{ParseArcs}},
\code{\link{ParseQuery}}, \code{\link{PlotSD}}
}
