% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2_buildSOM.R
\name{PeaksAndValleys}
\alias{PeaksAndValleys}
\title{Find peaks and valleys in one-dimensional data}
\usage{
PeaksAndValleys(data, minDensityThreshold = 0.05, ...)
}
\arguments{
\item{data}{array containing the data points}

\item{minDensityThreshold}{Only counts peaks which density > threshold}

\item{...}{Other parameters to be passed on to density}
}
\value{
A list containing the density result, peaks and valleys
}
\description{
Find peaks and valleys in one-dimensional data
}

