% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{Plot1DScatters}
\alias{Plot1DScatters}
\title{Plot1DScatters}
\usage{
Plot1DScatters(
  fsom,
  channels = fsom$map$colsUsed,
  clusters = NULL,
  metaclusters = NULL,
  maxBgPoints = 3000,
  sizeBgPoints = 0.5,
  colBgPoints = "gray",
  maxPoints = 1000,
  sizePoints = 0.5,
  yLim = NULL,
  xLabels = c("marker"),
  centers = FALSE,
  colors = NULL,
  ncol = NULL,
  nrow = NULL,
  width = NULL,
  height = NULL,
  plotFile = "1DScatterPlots.png"
)
}
\arguments{
\item{fsom}{FlowSOM object, as created by \code{\link{FlowSOM}}}

\item{channels}{Vector in which each element is a channel
or a marker}

\item{clusters}{Vector with indices of clusters of interest}

\item{metaclusters}{Vector with indices of metaclusters of interest}

\item{maxBgPoints}{Maximum number of background cells to plot}

\item{sizeBgPoints}{Size of the background cells}

\item{colBgPoints}{Color of background cells}

\item{maxPoints}{Maximum number of (meta)cluster cells to plot}

\item{sizePoints}{Size of the (meta)cluster cells}

\item{yLim}{Optional vector of a lower and upper limit of the y-axis}

\item{xLabels}{Determines the label of the x-axis. Can be 
"marker" and\\or "channel" or abbreviations.
 Default = "marker".}

\item{centers}{Add cluster centers to plot. Default = FALSE.}

\item{colors}{Vector of colors for all the cells in the selected nodes. 
First the clusters are colored, then the metaclusters. 
If \code{NULL}, the default ggplot colors, are used.}

\item{ncol}{Number of columns in the final plot, optional}

\item{nrow}{Number of rows in the final plot, optional}

\item{width}{Width of png file. By default NULL the width parameter is 
estimated based on the input.}

\item{height}{Height of png file. By default NULL the width parameter is 
estimated based on the input.}

\item{plotFile}{If a filepath for a png is given (default = 
1DScatterPlots.png), the plots will be plotted in 
the corresponding png file. If \code{NULL}, a list 
of ggplot objects will be returned}
}
\value{
If \code{plot} is \code{TRUE}, nothing is returned and a plot is 
        drawn in which background cells are plotted in gray and the cells of
        the selected nodes in color. If \code{plot} is \code{FALSE}, a ggplot
         objects list is returned.
}
\description{
Function to draw 1D scatter plots of FlowSOM (meta)clusters
}
\details{
Plot multiple 1D scatter plots in a png file. A subset of fsom$data is
plotted in gray, and those of the selected clusters and metaclusters are
plotted in color.
}
\examples{

# Identify the files
fcs <- flowCore::read.FCS(system.file("extdata", "68983.fcs",
                                     package = "FlowSOM"))
# Build a FlowSOM object
flowSOM.res <- FlowSOM(fcs,
                      scale = TRUE,
                      compensate = TRUE,
                      transform = TRUE,
                      toTransform = 8:18,
                      colsToUse = c(9, 12, 14:18),
                      nClus = 10,
                      seed = 1)

# Make the 1D scatter plots of the clusters and metaclusters of interest
Plot1DScatters(fsom = flowSOM.res,
              channels = c("PE-Cy7-A", "PE-Cy5-A"),
              clusters = c(1, 48, 49, 82, 95),
              metaclusters = list(c(1, 4), 9),
              xLabels = c("marker", "channel"),
              plotFile = tempfile(fileext = ".png"))
Plot1DScatters(flowSOM.res, 
              metaclusters = 4, 
              clusters = c(1, 2),
              colors = c("red", "green", "blue"),
              plotFile = tempfile(fileext = ".png"))                       

}
