% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotClusters2D}
\alias{PlotClusters2D}
\title{Plot nodes on scatter plot}
\usage{
PlotClusters2D(fsom, marker1, marker2, nodes, main = "", col = "#FF0000",
  maxBgPoints = NULL, ...)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{marker1}{Marker to plot on the x-axis}

\item{marker2}{Marker to plot on the y-axis}

\item{nodes}{Nodes of which the cells should be plotted in red}

\item{main}{Title of the plot}

\item{col}{Colors for all the cells in the selected nodes (ordered array)}

\item{maxBgPoints}{Maximum number of background points to plot}

\item{...}{Other parameters to pass on to plot}
}
\value{
Nothing is returned. A plot is drawn in which all cells are plotted
        in black and the cells of the selected nodes in red.
}
\description{
Plot a 2D scatter plot. All cells of fsom$data are plotted in
black, and those of the selected nodes are plotted in red.
The nodes in the grid are indexed starting from the left bottom,
first going right, then up. E.g. In a 10x10 grid, the node at
top left will have index 91.
}
\examples{
# Read from file, build self-organizing map and minimal spanning tree
   fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
   flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                            scale=TRUE)
   flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
  flowSOM.res <- BuildMST(flowSOM.res)

   # Plot cells
   PlotClusters2D(flowSOM.res,1,2,91)
}
\seealso{
\code{\link{PlotNumbers}},\code{\link{PlotCenters}},
         \code{\link{BuildMST}}
}

