% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotGroups}
\alias{PlotGroups}
\title{Plot differences between groups}
\usage{
PlotGroups(fsom, groups, tresh = NULL, p_tresh = 0.05, heatmap = FALSE, ...)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}} or
the first list item of \code{\link{FlowSOM}}}

\item{groups}{groups result as generated by \code{\link{CountGroups}}}

\item{tresh}{Relative difference in groups before the node is coloured}

\item{p_tresh}{Threshold on p-value from wilcox-test before the node is
coloured. If this is not NULL, tresh will be ignored.}

\item{heatmap}{If TRUE, the scores are plotted in a gradient instead
of only the selection that passes the threshold}

\item{...}{Other parameters to pass to \code{\link{PlotStars}}}
}
\value{
A vector containing the labels assigned to the nodes for
        all groups except the first
}
\description{
Plot FlowSOM trees, where each node is represented by 
a star chart indicating mean marker values, the size of the node is relative
to the mean percentage of cells present in each
}
\examples{
   ## Use the wrapper function to build a flowSOM object (saved in fsom[[1]])
   ## and a metaclustering (saved in fsom[[2]])
   # fsom <- FlowSOM(ff,compensate = FALSE, transform = FALSE,scale = TRUE,
   #              colsToUse = colsToUse, nClus = 10, silent = FALSE,
   #             xdim=7, ydim=7)
   
   ## Make a list with for each group a list of files
   ## The reference group should be the first
   # groups <- list("C"=file.path(workDir,grep("C",files,value = TRUE)),
   #             "D"=file.path(workDir,grep("D",files,value=TRUE)))
   
   ## Compute the percentages for all groups
   # groups_res <- CountGroups(fsom[[1]],groups)
  
   ## Plot the groups. For all groups except the first, differences with the
   ## first group are indicated
   # annotation <- PlotGroups(fsom[[1]],groups_res)
   
}
\seealso{
\code{\link{PlotStars}},\code{\link{CountGroups}}
}
