% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotLabels}
\alias{PlotLabels}
\title{Plot a label in each node}
\usage{
PlotLabels(fsom, labels, view = "MST", main = NULL,
  nodeSize = fsom$MST$size, fontSize = 1, backgroundValues = NULL,
  backgroundColor = function(n) {     grDevices::rainbow(n, alpha = 0.3)
  }, backgroundLim = NULL, backgroundBreaks = NULL)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{labels}{A label for every node}

\item{view}{Preferred view, options: "MST", "grid" or "tSNE" (if this
option was selected while building the MST)}

\item{main}{Title of the plot}

\item{nodeSize}{Nodesize. The plot might be easier to read if this is a 
constant number, e.g. 10 or 15}

\item{fontSize}{Fontsize, passed to label.cex}

\item{backgroundValues}{Values to be used for background coloring, either
numerical values or something that can be made into
a factor (e.g. a clustering)}

\item{backgroundColor}{Colorpalette to be used for the background coloring
. Can be either a function or an array specifying
colors}

\item{backgroundLim}{Only used when backgroundValues are numerical. 
Defaults to min and max of the backgroundValues.}

\item{backgroundBreaks}{Breaks to pass on to \code{\link{cut}}, to split
numerical background values. If NULL, the length of
backgroundColor will be used (default 100).}
}
\value{
Nothing is returned. A plot is drawn in which each node 
        is assigned a label
}
\description{
Plot FlowSOM grid or tree, with in each node a label. Especially useful to
show metacluster numbers
}
\examples{
# Read from file, build self-organizing map and minimal spanning tree
fileName <- system.file("extdata", "68983.fcs", package="FlowSOM")
ff <- flowCore::read.FCS(fileName)
ff <- flowCore::compensate(ff, ff@description$SPILL)
ff <- flowCore::transform(ff, flowCore::estimateLogicle(ff,
                                               flowCore::colnames(ff)[8:18]))
flowSOM.res <- FlowSOM(ff,
                       scale=TRUE,
                       colsToUse=c(9,12,14:18),
                       nClus = 10,
                       seed = 1)

# Plot the node IDs
PlotLabels( flowSOM.res$FlowSOM, flowSOM.res$metaclustering, nodeSize=15)

}
\seealso{
\code{\link{PlotNumbers}}
}
