% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{PlotMFIHeatmap}
\alias{PlotMFIHeatmap}
\title{PlotMFIHeatmap}
\usage{
PlotMFIHeatmap(
  fsom,
  level = c("clusters", "metaclusters"),
  heatmapColors = c("#313695", "#4575B4", "#74ADD1", "#ABD9E9", "#E0F3F8", "#FEE090",
    "#FDAE61", "#F46D43", "#D73027", "#A50026"),
  annotationColors = gg_color_hue,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  labels = FALSE
)
}
\arguments{
\item{fsom}{FlowSOM object, as created by \code{\link{FlowSOM}}}

\item{level}{A vector with "clusters" and/or "metaclusters (or abbreviations)
to show clusters and/or metaclusters. Default shows both.}

\item{heatmapColors}{Vector of colors for the heatmap. The default colors are
RdYlBu}

\item{annotationColors}{A color function to color the metacluster column 
annotation. The default colors are the ggplot colors}

\item{cluster_rows}{Whether or not to cluster the rows. Default = FALSE}

\item{cluster_columns}{Whether or not to cluster the columns Default = FALSE}

\item{labels}{Whether or not to display the MFIs. Default = FALSE}
}
\value{
Plot
}
\description{
Plots a heatmap of the cluster medians and metacluster medians
}
\examples{

# Identify the files
fcs <- flowCore::read.FCS(system.file("extdata", "68983.fcs", 
                                      package = "FlowSOM"))
# Build a FlowSOM object
flowSOM.res <- FlowSOM(fcs, 
                       scale = TRUE,
                       compensate = TRUE, 
                       transform = TRUE,
                       toTransform = 8:18, 
                       colsToUse = c(9, 12, 14:18),
                       nClus = 10,
                       seed = 1)
p <- PlotMFIHeatmap(flowSOM.res)                       


}
