% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotNumbers}
\alias{PlotNumbers}
\title{Plot the index of each node}
\usage{
PlotNumbers(fsom, view = "MST", main = NULL)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{view}{Preferred view, options: "MST", "grid" or "tSNE" (if this
option was selected while building the MST)}

\item{main}{Title of the plot}
}
\value{
Nothing is returned. A plot is drawn in which each node
        is assigned a number
}
\description{
Plot FlowSOM grid or tree, with in each node a number indicating it's index
}
\examples{
# Read from file, build self-organizing map and minimal spanning tree
fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                         scale=TRUE)
flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
flowSOM.res <- BuildMST(flowSOM.res)

# Plot some random values
PlotNumbers(flowSOM.res)
}
\seealso{
\code{\link{PlotMarker}},\code{\link{PlotStars}},
         \code{\link{PlotPies}},\code{\link{PlotCenters}},
         \code{\link{BuildMST}}
}

