% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotPies}
\alias{PlotPies}
\title{Plot comparison with other clustering}
\usage{
PlotPies(fsom, cellTypes, view = "MST",
  colorPalette = colorRampPalette(c("white", "#00007F", "blue", "#007FFF",
  "cyan", "#7FFF7F", "yellow", "#FF7F00", "red")), backgroundValues = NULL,
  backgroundColor = function(n) {     rainbow(n, alpha = 0.3) },
  backgroundBreaks = NULL, legend = TRUE, main = "")
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{cellTypes}{Array of factors indicating the celltypes}

\item{view}{Preferred view, options: "MST", "grid" or "tSNE" (if this
option was selected while building the MST)}

\item{colorPalette}{Colorpalette to be used for the markers}

\item{backgroundValues}{Values to be used for background coloring, either
numerical values or something that can be made into
a factor (e.g. a clustering)}

\item{backgroundColor}{Colorpalette to be used for the background coloring
. Can be either a function or an array specifying
colors}

\item{backgroundBreaks}{Breaks to pass on to \code{\link{cut}}, to split
numerical background values. If NULL, the length of
backgroundColor will be used (default 100).}

\item{legend}{Logicle, if T add a legend}

\item{main}{Title of the plot}
}
\value{
Nothing is returned. A plot is drawn in which each node is
represented by a pie chart indicating the percentage of cells present
of each cell type. At the end, the layout is set to 1 figure again.
}
\description{
Plot FlowSOM grid or tree, with pies indicating another clustering
   or manual gating result
}
\examples{
# Read from file, build self-organizing map and minimal spanning tree
   fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
   flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                            scale=TRUE)
   flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
   flowSOM.res <- BuildMST(flowSOM.res)

   # Get the manually gated labels using a gatingML file
   # and the flowUtils library
    ff <- read.FCS(fileName)
   ff_c <- compensate(ff,ff@description$SPILL)
   colnames(ff_c)[8:18] <- paste("Comp-",colnames(ff_c)[8:18],sep="")
   library(flowUtils)
   flowEnv <- new.env()
   gatingFile <- system.file("extdata","manualGating.xml",
                             package="FlowSOM")
   read.gatingML(gatingFile, flowEnv)
   filterList <- list( "B cells" = flowEnv$ID52300206,
                       "ab T cells" = flowEnv$ID785879196,
                       "yd T cells" = flowEnv$ID188379411,
                       "NK cells" = flowEnv$ID1229333490,
                       "NKT cells" = flowEnv$ID275096433
   )
   results <- list()
   for(cellType in names(filterList)){
       results[[cellType]] <- filter(ff_c,filterList[[cellType]])@subSet
   }
   manual <- rep("Unknown",nrow(ff))
   for(celltype in names(results)){
       manual[results[[celltype]]] <- celltype
   }

   # Plot pies indicating the percentage of cell types present in the nodes
   PlotPies(flowSOM.res,manual)
}
\seealso{
\code{\link{PlotStars}},\code{\link{PlotMarker}},
\code{\link{PlotCenters}},\code{\link{BuildMST}}
}

