% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{plotStarLegend}
\alias{plotStarLegend}
\title{Plot legend for star plot}
\usage{
plotStarLegend(labels, colors = grDevices::rainbow(length(labels)), main = "")
}
\arguments{
\item{labels}{Names to show in the legend}

\item{colors}{Corresponding colors}

\item{main}{Title of the legend}
}
\value{
Nothing is returned. A plot is drawn with 1 star chart, which is 
filled completely and annotated with the given labels.
}
\description{
Plot a single star chart, annotated with labels
}
\seealso{
\code{\link{PlotStars}}
}
