% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotStars}
\alias{PlotStars}
\title{Plot star charts}
\usage{
PlotStars(fsom, markers = fsom$map$colsUsed, view = "MST",
  colorPalette = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan",
  "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000")), backgroundValues = NULL,
  backgroundColor = function(n) {     rainbow(n, alpha = 0.3) },
  legend = TRUE, query = NULL, main = "")
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{markers}{Array of markers to use. Default: the markers used to build
the tree}

\item{view}{Preferred view, options: "MST", "grid" or "tSNE" (if this
option was selected while building the MST)}

\item{colorPalette}{Colorpalette to be used for the markers}

\item{backgroundValues}{Values to be used for background coloring, either
numerical values or something that can be made into
a factor (e.g. a clustering)}

\item{backgroundColor}{Colorpalette to be used for the background coloring
. Can be either a function or an array specifying
colors}

\item{legend}{Logicle, if T add a legend}

\item{query}{Show a low/high profile for certain markers in the legend.
See also \code{\link{QueryStarPlot}}}

\item{main}{Title of the plot}
}
\value{
Nothing is returned. A plot is drawn in which each node is
represented by a star chart indicating the median fluorescence intensities.
Resets the layout back to 1 plot at the end.
}
\description{
Plot FlowSOM grid or tree, where each node is represented by
a star chart indicating mean marker values
}
\examples{
# Read from file, build self-organizing map and minimal spanning tree
   fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
   flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                            scale=TRUE)
   flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
   flowSOM.res <- BuildMST(flowSOM.res)

   # Plot stars indicating the MFI of the cells present in the nodes
   PlotStars(flowSOM.res)
}
\seealso{
\code{\link{PlotPies}},\code{\link{PlotMarker}},
\code{\link{PlotCenters}}, \code{\link{BuildMST}}
}

