% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotVariable}
\alias{PlotVariable}
\title{Plot a variable for all nodes}
\usage{
PlotVariable(fsom, variable, view = "MST", main = NULL,
  colorPalette = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan",
  "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000")))
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{variable}{Vector containing a value for each node}

\item{view}{Preferred view, options: "MST", "grid" or "tSNE" (if this
option was selected while building the MST)}

\item{main}{Title of the plot}

\item{colorPalette}{Color palette to use}
}
\value{
Nothing is returned. A plot is drawn in which each node
        is coloured depending on its value for the given variable
}
\description{
Plot FlowSOM grid or tree, coloured by node values given in variable
}
\examples{
# Read from file, build self-organizing map and minimal spanning tree
fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                         scale=TRUE)
flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
flowSOM.res <- BuildMST(flowSOM.res)

# Plot some random values
rand <- runif(flowSOM.res$map$nNodes)
PlotVariable(flowSOM.res,rand)
}
\seealso{
\code{\link{PlotMarker}},\code{\link{PlotStars}},
         \code{\link{PlotPies}},\code{\link{PlotCenters}},
         \code{\link{BuildMST}}
}

