% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_FlowSOM.R
\name{print.FlowSOM}
\alias{print.FlowSOM}
\title{Print FlowSOM object}
\usage{
\method{print}{FlowSOM}(x, ...)
}
\arguments{
\item{x}{FlowSOM object to print information about}

\item{...}{Further arguments, not used}
}
\description{
Print FlowSOM object
}
\examples{
fileName <- system.file("extdata", "68983.fcs", package = "FlowSOM")
flowSOM.res <- FlowSOM(fileName, compensate = TRUE, transform = TRUE,
                      scale = TRUE, colsToUse = c(9, 12, 14:18), nClus = 10)
print(flowSOM.res)

}
