% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_FlowSOM.R
\name{ProcessGatingML}
\alias{ProcessGatingML}
\title{Process a gatingML file}
\usage{
ProcessGatingML(flowFrame, gatingFile, gateIDs, cellTypes, silent = FALSE)
}
\arguments{
\item{flowFrame}{The flowFrame to apply the gating on}

\item{gatingFile}{The gatingML file to read}

\item{gateIDs}{Named vector containing ids to extract from the
gatingML file to use in the matrix}

\item{cellTypes}{Cell types to use for labeling the cells. Should be a
subset of the names of the gateIDs}

\item{silent}{If FALSE, show messages of which gates are being
processed}
}
\value{
This function returns a list in which the first element ("matrix")
is a matrix containing filtering results for each specified gate and the
second element ("manual") is a vector which assigns a label to each cell
}
\description{
Reads a gatingML file using the \code{\link{flowUtils}} library and
returns a list with a matrix containing filtering results for each specified
gate and a vector with a label for each cell
}
\examples{

   # Read the flowFrame
   fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
   ff <- flowCore::read.FCS(fileName)
   ff_c <- flowCore::compensate(ff,flowCore::description(ff)$SPILL)
   flowCore::colnames(ff_c)[8:18] <- paste("Comp-",
                                     flowCore::colnames(ff_c)[8:18],
                                     sep="")

   # Specify the gating file and the gates of interest
   gatingFile <- system.file("extdata","manualGating.xml",
                             package="FlowSOM")
   gateIDs <- c( "B cells"=8,
                 "ab T cells"=10,
                 "yd T cells"=15,
                 "NK cells"=5,
                 "NKT cells"=6)
   cellTypes <- c("B cells","ab T cells","yd T cells",
                 "NK cells","NKT cells")
   gatingResult <- ProcessGatingML(ff_c, gatingFile, gateIDs, cellTypes)


   # Build a FlowSOM tree
   flowSOM.res <- FlowSOM(ff_c,compensate=FALSE,transform=TRUE,
                         toTransform=8:18,colsToUse=c(9,12,14:18),nClus=10)
   # Plot pies indicating the percentage of cell types present in the nodes
   PlotPies(flowSOM.res[[1]],gatingResult$manual)

}
\seealso{
\code{\link{PlotPies}}
}
