% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{QueryStarPlot}
\alias{QueryStarPlot}
\title{Query a certain cell type}
\usage{
QueryStarPlot(fsom, query, plot = TRUE, color = "#ca0020", debug = FALSE, ...)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}} or
the first list item of \code{\link{FlowSOM}}}

\item{query}{Array containing "high" or "low" for the specified column
names of the FlowSOM data}

\item{plot}{If true, a plot with a gradient of scores for the nodes is 
shown}

\item{color}{Color to use for nodes with a high score in the plot}

\item{debug}{If TRUE, some extra output will be printed}

\item{...}{Other parameters to pass to \code{\link{PlotStars}}}
}
\value{
A list, containing the ids of the selected nodes, the individual
        scores for all nodes and the scores for each marker for each node
}
\description{
Identify nodes in the tree which resemble a certain profile of "high"
or "low" marker expressions.
}
\examples{
   file <- system.file("extdata", "68983.fcs", package="FlowSOM")
   # Use the wrapper function to build a flowSOM object (saved in fsom[[1]])
   # and a metaclustering (saved in fsom[[2]])
   fsom <- FlowSOM(file,compensate = TRUE, transform = TRUE,scale = TRUE,
                  colsToUse = c(9,12,14:18), nClus = 10, silent = FALSE,
                  xdim=7, ydim=7)
   query <- c("PE-Cy7-A" = "high", #CD3
              "APC-Cy7-A" = "high", #TCRb
              "Pacific Blue-A" = "high") #CD8
   query_res <- QueryStarPlot(UpdateNodeSize(fsom[[1]],reset=TRUE), query)
   
   cellTypes <- factor(rep("Unknown",49),levels=c("Unknown","CD8 T cells"))
   cellTypes[query_res$selected] <- "CD8 T cells"
   PlotStars(fsom[[1]],
                 backgroundValues=cellTypes,
                 backgroundColor=c("#FFFFFF00","#ca0020aa"))
   
}
