% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_metaClustering.R
\name{ReassignMetaclusters}
\alias{ReassignMetaclusters}
\title{ReassignMetaclusters}
\usage{
ReassignMetaclusters(fsom, metaclustering)
}
\arguments{
\item{fsom}{Result of calling the FlowSOM function}

\item{metaclustering}{Vector with the metacluster names for all clusters}
}
\value{
Updated FlowSOM object
}
\description{
Adapt the metaclustering. Can be used to either split up metaclusters,
or potentially merge some metaclusters.
}
\examples{
fileName <- system.file("extdata", "68983.fcs", package = "FlowSOM")
ff <- flowCore::read.FCS(fileName)
ff <- flowCore::compensate(ff, flowCore::keyword(ff)[["SPILL"]])
ff <- flowCore::transform(ff,
         flowCore::transformList(colnames(flowCore::keyword(ff)[["SPILL"]]),
                                flowCore::logicleTransform()))
flowSOM.res <- FlowSOM(ff,
                       scale = TRUE,
                       colsToUse = c(9, 12, 14:18), 
                       nClus = 5,
                       seed = 1)
                       
PlotStars(flowSOM.res, backgroundValues = flowSOM.res$metaclustering)

# Split up metacluster 3
MC_or <- flowSOM.res$metaclustering
MC_new <- c(MC_or)
MC_new[c(81:86, 91:96)] <- "5b"

flowSOM.res <- ReassignMetaclusters(flowSOM.res, MC_new)
PlotStars(flowSOM.res, backgroundValues = flowSOM.res$metaclustering)
PlotNumbers(flowSOM.res, level = "metaclusters")
GetCounts(flowSOM.res)

}
