% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_metaClustering.R
\name{ReorderMetaclusters}
\alias{ReorderMetaclusters}
\title{ReorderMetaclusters}
\usage{
ReorderMetaclusters(fsom, order)
}
\arguments{
\item{fsom}{Result of calling the FlowSOM function}

\item{order}{Prefered order of the metaclusters}
}
\value{
Updated FlowSOM object
}
\description{
Changes the order of the metaclusters in the FlowSOM object
}
\examples{
fileName <- system.file("extdata", "68983.fcs", package = "FlowSOM")
ff <- flowCore::read.FCS(fileName)
ff <- flowCore::compensate(ff, flowCore::keyword(ff)[["SPILL"]])
ff <- flowCore::transform(ff,
         flowCore::transformList(colnames(flowCore::keyword(ff)[["SPILL"]]),
                                flowCore::logicleTransform()))
flowSOM.res <- FlowSOM(ff,
                       scale = TRUE,
                       colsToUse = c(9, 12, 14:18), 
                       nClus = 5,
                       seed = 1)
                       
PlotStars(flowSOM.res, backgroundValues = flowSOM.res$metaclustering)

flowSOM.res <- ReorderMetaclusters(flowSOM.res, c(3, 2, 5, 1, 4))
PlotStars(flowSOM.res, backgroundValues = flowSOM.res$metaclustering)

}
