% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{ScaleStarHeights}
\alias{ScaleStarHeights}
\title{ScaleStarHeights}
\usage{
ScaleStarHeights(data, nodeSizes)
}
\arguments{
\item{data}{MedianValues of relevant markers extraced from 
FlowSOM object}

\item{nodeSizes}{A vector that is returned from \code{ParseNodeSize}}
}
\value{
A dataframe consisting of the scaled values of the stars. 
        The stars are scaled between 0 and the maximum of all stars
}
\description{
Scales starheights
}
\details{
Function that scales the star values between 0 and the node size
}
\seealso{
\code{\link{PlotFlowSOM}}, \code{\link{ParseNodeSize}}, 
\code{\link{AutoMaxNodeSize}}
}
