% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{shiftFunction}
\alias{shiftFunction}
\title{shiftFunction}
\usage{
shiftFunction(x, n)
}
\arguments{
\item{x}{Vector to use}

\item{n}{Number of places to shift}
}
\value{
Shifted vector
}
\description{
shiftFunction
}
