% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{UpdateFlowSOM}
\alias{UpdateFlowSOM}
\title{UpdateFlowSOM}
\usage{
UpdateFlowSOM(fsom)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}} or
\code{\link{FlowSOM}}}
}
\value{
A FlowSOM object
}
\description{
Update old FlowSOM object to a new one and checks if it is a flowSOM object
}
\details{
Determines whether or not the fsom input is of class "FlowSOM" and returns  
the FlowSOM object and metaclustering object inside fsom
}
\seealso{
\code{\link{PlotFlowSOM}}
}
