\name{UpdateNodeSize}
\alias{UpdateNodeSize}
\title{Update nodesize of FlowSOM object}
\description{Add size property to the graph based on cellcount for each node}
\usage{UpdateNodeSize(fsom, reset=FALSE)}
\arguments{
    \item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}
    \item{reset}{logical. If \code{TRUE}, all nodes get the same size}
}
\value{Updated FlowSOM object}
\seealso{\code{\link{BuildMST}}}
\examples{
    # Read from file, build self-organizing map and minimal spanning tree
    fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
    flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                            scale=TRUE)
    flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
    flowSOM.res <- BuildMST(flowSOM.res)
    
    # Give all nodes same size
    flowSOM.res <- UpdateNodeSize(flowSOM.res,reset=TRUE)
    PlotStars(flowSOM.res)
    
    # Node sizes relative to amount of cells assigned to the node
    flowSOM.res <- UpdateNodeSize(flowSOM.res)
    PlotStars(flowSOM.res)
}