\name{CountGroups}
\alias{CountGroups}
\title{Compute differences between groups}
\description{Map new data to the flowSOM grid and average over all files from
one group}
\usage{CountGroups(fsom, groups, plot=TRUE, silent=FALSE)}
\arguments{
    \item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}
    \item{groups}{List with for every group a vector with filepaths}
    \item{plot}{Logical: if \code{TRUE}, the star plot is shown for each
    individual file, with adapted node size}
    \item{silent}{Logical: if \code{FALSE}, progress is printed}
}
\value{A list including the counts, percentages, means and medians}
\seealso{
    \code{\link{PlotStars}},\code{\link{PlotGroups}},code{\link{NewData}}
}
\examples{
    ## Use the wrapper function to build a flowSOM object (saved in fsom[[1]])
    ## and a metaclustering (saved in fsom[[2]])
    # fsom <- FlowSOM(ff,compensate = FALSE, transform = FALSE,scale = TRUE,
    #               colsToUse = colsToUse, nClus = 10, silent = FALSE,
    #               xdim=7, ydim=7)

    ## Make a list with for each group a list of files
    ## The reference group should be the first
    #groups <- list("C"=file.path(workDir,grep("C",files,value = TRUE)),
    #          "D"=file.path(workDir,grep("D",files,value=TRUE)))

    ## Compute the percentages for all groups
    # groups_res <- CountGroups(fsom[[1]],groups)

    ## Plot the groups. For all groups except the first, differences with the
    ## first group are indicated
    # annotation <- PlotGroups(fsom[[1]],groups_res)
}