% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_deprecated.R
\name{GetCVs}
\alias{GetCVs}
\title{Get CV values for all clusters}
\usage{
GetCVs(fsom)
}
\arguments{
\item{fsom}{FlowSOM object as generated by the FlowSOM function
or the BuildSOM function}
}
\value{
Matrix with coefficient of variation values for each marker

fileName <- system.file("extdata", "68983.fcs", package="FlowSOM")
flowSOM.res <- FlowSOM(fileName, compensate=TRUE,transform=TRUE,
                      scale=TRUE,colsToUse=c(9,12,14:18),nClus=10)
cvs <- GetClusterCVs(flowSOM.res)
}
\description{
Get CV values for all clusters
}
