% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_buildSOM.R
\name{GetMFIs}
\alias{GetMFIs}
\title{Get MFI values for all clusters}
\usage{
GetMFIs(fsom)
}
\arguments{
\item{fsom}{FlowSOM object as generated by the FlowSOM function
or the BuildSOM function}
}
\value{
Matrix with median values for each marker
}
\description{
Get MFI values for all clusters
}
