% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_buildSOM.R
\name{Initialize_PCA}
\alias{Initialize_PCA}
\title{Create a grid from first 2 PCA components}
\usage{
Initialize_PCA(data, xdim, ydim)
}
\arguments{
\item{data}{matrix in which each row represents a point}

\item{xdim}{x dimension of the grid}

\item{ydim}{y dimension of the grid}
}
\value{
array containing the selected selected rows
}
\description{
Create a grid from first 2 PCA components
}
\examples{

points <- matrix(1:1000, ncol = 10)
selection <- Initialize_PCA(points, 3, 3)

}
