% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{Plot1DScatters}
\alias{Plot1DScatters}
\title{Plot1DScatters}
\usage{
Plot1DScatters(
  fsom,
  channels = fsom$map$colsUsed,
  clusters = NULL,
  metaclusters = NULL,
  maxBgPoints = 3000,
  sizeBgPoints = 0.5,
  colBgPoints = "gray",
  maxPoints = 1000,
  sizePoints = 0.5,
  yLim = NULL,
  xLabels = c("marker"),
  centers = FALSE,
  colors = NULL,
  plotFile = NULL
)
}
\arguments{
\item{fsom}{FlowSOM object, as created by \code{\link{FlowSOM}}}

\item{channels}{Vector in which each element is a channel
or a marker}

\item{clusters}{Vector with indices of clusters of interest}

\item{metaclusters}{Vector with indices of metaclusters of interest}

\item{maxBgPoints}{Maximum number of background cells to plot}

\item{sizeBgPoints}{Size of the background cells}

\item{colBgPoints}{Color of background cells}

\item{maxPoints}{Maximum number of (meta)cluster cells to plot}

\item{sizePoints}{Size of the (meta)cluster cells}

\item{yLim}{Optional vector of a lower and upper limit of the y-axis}

\item{xLabels}{Determines the label of the x-axis. Can be 
"marker" and\\or "channel" or abbrevations.
 Default = "marker".}

\item{centers}{Add cluster centers to plot. Default = FALSE.}

\item{colors}{Vector of colors for all the cells in the selected nodes. 
First the clusters are colored, then the metaclusters. 
If \code{NULL}, the default ggplot colors, are used.}

\item{plotFile}{If a filepath for graphic device is given (default = NULL), 
the plots will be plotted in the corresponding file. 
If \code{NULL}, a ggplot object will be returned}
}
\value{
Plot
}
\description{
Plot an overview of the clusters or metaclusters per marker with all the cells
as reference
}
\examples{

# Identify the files
fcs <- flowCore::read.FCS(system.file("extdata", "68983.fcs", 
                                      package = "FlowSOM"))
# Build a FlowSOM object
flowSOM.res <- FlowSOM(fcs, 
                       scale = TRUE,
                       compensate = TRUE, 
                       transform = TRUE,
                       toTransform = 8:18, 
                       colsToUse = c(9, 12, 14:18),
                       nClus = 10,
                       seed = 1)
p <- Plot1DScatters(flowSOM.res, metaclusters = 4, clusters = c(1, 2))                       

}
