% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotMarker}
\alias{PlotMarker}
\title{Plot marker values}
\usage{
PlotMarker(fsom, marker = NULL, view = TRUE, main = NULL,
  colorPalette = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan",
  "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000")))
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{marker}{Name or index of marker to plot}

\item{view}{Preferred view, options: "MST", "grid" or "tSNE" (if this
option was selected while building the MST)}

\item{main}{Title of the plot}

\item{colorPalette}{Color palette to use}
}
\value{
Nothing is returned. A plot is drawn in which each node 
        is coloured depending on its median value for the given marker
}
\description{
Plot FlowSOM grid or tree, coloured by node values for a specific marker
}
\examples{
# Read from file, build self-organizing map and minimal spanning tree
fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                         scale=TRUE)
flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
flowSOM.res <- BuildMST(flowSOM.res)

# Plot one marker
PlotMarker(flowSOM.res,"FSC-A")

}
\references{
This visualization technique resembles SPADE results. 
M. Linderman, P. Qiu, E. Simonds and Z. Bjornson (). spade: SPADE -- An 
analysis and visualization tool for Flow Cytometry. R package version 1.12.2.
http://cytospade.org
}
\seealso{
\code{\link{PlotStars}},\code{\link{PlotPies}},
         \code{\link{PlotCenters}},\code{\link{BuildMST}}
}

