% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotNumbers}
\alias{PlotNumbers}
\title{Plot the index of each node}
\usage{
PlotNumbers(
  fsom,
  view = "MST",
  main = NULL,
  nodeSize = fsom$MST$size,
  fontSize = 1,
  backgroundValues = NULL,
  backgroundColor = function(n) {     grDevices::rainbow(n, alpha = 0.3) },
  backgroundLim = NULL,
  backgroundBreaks = NULL
)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{view}{Preferred view, options: "MST", "grid" or "tSNE" (if this
option was selected while building the MST)}

\item{main}{Title of the plot}

\item{nodeSize}{Nodesize. The plot might be easier to read if this is a 
constant number, e.g. 10 or 15}

\item{fontSize}{Fontsize, passed to label.cex}

\item{backgroundValues}{Values to be used for background coloring, either
numerical values or something that can be made into
a factor (e.g. a clustering)}

\item{backgroundColor}{Colorpalette to be used for the background coloring
. Can be either a function or an array specifying
colors}

\item{backgroundLim}{Only used when backgroundValues are numerical. 
Defaults to min and max of the backgroundValues.}

\item{backgroundBreaks}{Breaks to pass on to \code{\link{cut}}, to split
numerical background values. If NULL, the length of
backgroundColor will be used (default 100).}
}
\value{
Nothing is returned. A plot is drawn in which each node 
        is assigned a number
}
\description{
Plot FlowSOM grid or tree, with in each node a number indicating its index
}
\examples{
# Read from file, build self-organizing map and minimal spanning tree
fileName <- system.file("extdata", "68983.fcs", package="FlowSOM")
flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                         scale=TRUE)
flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
flowSOM.res <- BuildMST(flowSOM.res)

# Plot the node IDs
PlotNumbers(flowSOM.res)

# Adapt node size for easier readability
PlotNumbers(flowSOM.res, nodeSize=14)

}
\seealso{
\code{\link{PlotMarker}},\code{\link{PlotStars}},
         \code{\link{PlotPies}},\code{\link{PlotCenters}},
         \code{\link{BuildMST}}
}
