\name{PlotPies}
\alias{PlotPies}
\title{Plot comparison with other clustering}
\description{Plot FlowSOM grid or tree, with pies indicating another clustering
or manual gating result}
\usage{PlotPies(fsom, cellTypes, MST=TRUE, legend=TRUE, clusters=NULL,main="",
                colorPalette=colorRampPalette(c("white","#00007F", "blue", 
                "#007FFF","cyan", "#7FFF7F", "yellow", "#FF7F00", "red")))}
\arguments{
    \item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}
    \item{cellTypes}{Array of factors indicating the celltypes}
    \item{MST}{logical. If \code{TRUE}, plot tree, else plot grid}
    \item{legend}{logical. Sometimes the position of the legend is not great,
    so it might be easier to plot without}
    \item{clusters}{optional, clustering of the SOM nodes}
    \item{main}{Title of the plot}
    \item{colorPalette}{Color palette to use}
}
\value{Nothing is returned. A plot is drawn in which each node is represented
by a pie chart indicating the percentage of cells present of each cell type}
\seealso{
    \code{\link{PlotStars}},\code{\link{PlotMarker}},\code{\link{PlotCenters}},
    \code{\link{BuildMST}}
}
\examples{
    # Read from file, build self-organizing map and minimal spanning tree
    fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
    flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                            scale=TRUE)
    flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
    flowSOM.res <- BuildMST(flowSOM.res)
    
    # Get the manually gated labels using a gatingML file 
    # and the flowUtils library
    ff <- read.FCS(fileName)
    ff_c <- compensate(ff,ff@description$SPILL)
    colnames(ff_c)[8:18] <- paste("Comp-",colnames(ff_c)[8:18],sep="")
    library(flowUtils)
    flowEnv <- new.env()
    gatingFile <- system.file("extdata","manualGating.xml", 
                            package="FlowSOM")
    read.gatingML(gatingFile, flowEnv) 
    filterList <- list( "B cells" = flowEnv$ID52300206,
                    "ab T cells" = flowEnv$ID785879196,
                    "yd T cells" = flowEnv$ID188379411,
                    "NK cells" = flowEnv$ID1229333490,
                    "NKT cells" = flowEnv$ID275096433
                    )
    results <- list()
    for(cellType in names(filterList)){
        results[[cellType]] <- filter(ff_c,filterList[[cellType]])@subSet
    }
    manual <- rep("Unknown",nrow(ff))
    for(celltype in names(results)){
        manual[results[[celltype]]] <- celltype
    }
    
    # Plot pies indicating the percentage of cell types present in the nodes
    PlotPies(flowSOM.res,manual)
}