% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotSD}
\alias{PlotSD}
\title{Plot SD}
\usage{
PlotSD(fsom, marker = NULL, view = "MST", main = NULL,
  colorPalette = grDevices::colorRampPalette(c("#00007F", "blue",
  "#007FFF", "cyan", "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000")),
  symmetric = FALSE, lim = NULL, backgroundValues = NULL,
  backgroundColor = function(n) {     grDevices::rainbow(n, alpha = 0.3)
  }, backgroundLim = NULL, backgroundBreaks = NULL)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{marker}{If a marker is given, the sd for this marker is shown.
Otherwise, the maximum ratio is used.}

\item{view}{Preferred view, options: "MST", "grid" or "tSNE" (if this
option was selected while building the MST)}

\item{main}{Title of the plot}

\item{colorPalette}{Color palette to use}

\item{symmetric}{Plot colours symmetric around zero}

\item{lim}{Variable limits}

\item{backgroundValues}{Values to be used for background coloring, either
numerical values or something that can be made into
a factor (e.g. a clustering)}

\item{backgroundColor}{Colorpalette to be used for the background coloring
. Can be either a function or an array specifying
colors}

\item{backgroundLim}{Only used when backgroundValues are numerical. 
Defaults to min and max of the backgroundValues.}

\item{backgroundBreaks}{Breaks to pass on to \code{\link{cut}}, to split
numerical background values. If NULL, the length of
backgroundColor will be used (default 100).}
}
\value{
Nothing is returned. A plot is drawn in which each node 
        is coloured depending on its standard deviation
}
\description{
--- Function in development, use with caution ---
Plot FlowSOM grid or tree, coloured by standard deviaton
}
\details{
From suggestion in email:
I am currently considering a way to summarize for each node all the SD as one value. 
After computing the SD matrix (nrow = # nodes, ncol = # markers), I compute the median value per column, 
then divide the SD matrix by it, and finally take the maximum ratio of each line (aka node). 
Doing so I got a unique dispersion score per node.
}
\examples{
# Read from file, build self-organizing map and minimal spanning tree
fileName <- system.file("extdata", "68983.fcs", package="FlowSOM")
flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                         scale=TRUE)
flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
flowSOM.res <- BuildMST(flowSOM.res)

PlotSD(flowSOM.res)

}
\seealso{
\code{\link{PlotMarker}},\code{\link{PlotStars}},
         \code{\link{PlotPies}},\code{\link{PlotCenters}},
         \code{\link{BuildMST}}
}
