% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{PlotStarLegend}
\alias{PlotStarLegend}
\title{PlotStarLegend}
\usage{
PlotStarLegend(markers, colors, starHeight = 1)
}
\arguments{
\item{markers}{Vector of markers used in legend}

\item{colors}{ColorPalette for the legend. Can be a vector or a 
function.}

\item{starHeight}{Star height. Default = 1.}
}
\value{
Returns nothing, but plots a legend for FlowSOM star plot
}
\description{
Plots star legend
}
\details{
Function makes the legend of the FlowSOM star plot
}
\examples{
PlotStarLegend(c("CD3", "CD4", "CD8"),
               FlowSOM_colors(3))

}
\seealso{
\code{\link{PlotFlowSOM}}
}
