\name{PlotStars}
\alias{PlotStars}
\title{Plot star charts}
\description{Plot FlowSOM grid or tree, where each node is represented by 
a star chart indicating mean marker values}
\usage{PlotStars(fsom, markers=fsom$map$colsUsed, MST=TRUE, legend=TRUE, 
                clusters=NULL,main="",colorPalette=colorRampPalette(
                c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow",
                "#FF7F00", "red", "#7F0000")),clusterColorPalette=
                function(n){rainbow(n,alpha=0.3)})}
\arguments{
    \item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}
    \item{markers}{Array of markers to use. Default: the markers used to build 
    the tree}
    \item{MST}{logical. If \code{TRUE}, plot tree, else plot grid}
    \item{legend}{logical. Sometimes the position of the legend is not great,
    so it might be easier to plot without}
    \item{clusters}{optional, clustering of the SOM nodes}
    \item{main}{Title of the plot}
    \item{colorPalette}{Colorpalette to be used}
    \item{clusterColorPalette}{Colorpalette to be used for the metaclusters}
}
\value{Nothing is returned. A plot is drawn in which each node is represented
by a star chart indicating the mean fluorescence intensities}
\seealso{
    \code{\link{PlotPies}},\code{\link{PlotMarker}},\code{\link{PlotCenters}},
    \code{\link{BuildMST}}
}
\examples{
    # Read from file, build self-organizing map and minimal spanning tree
    fileName <- system.file("extdata","lymphocytes.fcs",package="FlowSOM")
    flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                            scale=TRUE)
    flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
    flowSOM.res <- BuildMST(flowSOM.res)
    
    # Plot stars indicating the MFI of the cells present in the nodes
    PlotStars(flowSOM.res)
}