% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_buildMST.R
\name{PlotStars}
\alias{PlotStars}
\title{Plot star charts}
\usage{
PlotStars(
  fsom,
  markers = fsom$map$colsUsed,
  view = "MST",
  colorPalette = grDevices::colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan",
    "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000")),
  starBg = "white",
  backgroundValues = NULL,
  backgroundColor = function(n) {     grDevices::rainbow(n, alpha = 0.3) },
  backgroundLim = NULL,
  backgroundBreaks = NULL,
  backgroundSize = NULL,
  thresholds = NULL,
  legend = TRUE,
  query = NULL,
  range = "all",
  main = ""
)
}
\arguments{
\item{fsom}{FlowSOM object, as generated by \code{\link{BuildMST}}}

\item{markers}{Array of markers to use. Default: the markers used to build 
the tree}

\item{view}{Preferred view, options: "MST", "grid" or "tSNE" (if this
option was selected while building the MST)}

\item{colorPalette}{Colorpalette to be used for the markers}

\item{starBg}{Background color inside the star circle. Default is "white".
Can also be put to  "transparent" (as was the case for older 
versions).}

\item{backgroundValues}{Values to be used for background coloring, either
numerical values or something that can be made into
a factor (e.g. a clustering)}

\item{backgroundColor}{Colorpalette to be used for the background coloring
. Can be either a function or an array specifying
colors}

\item{backgroundLim}{Only used when backgroundValues are numerical. 
Defaults to min and max of the backgroundValues.}

\item{backgroundBreaks}{Breaks to pass on to \code{\link{cut}}, to split
numerical background values. If NULL, the length of
backgroundColor will be used (default 100).}

\item{backgroundSize}{Size of the background circles. Default 15.}

\item{thresholds}{Optional. Array containing a number for each of the 
markers to be used as the split between high/low. 
If provided, the percentage of positive cells is
indicated instead of the MFI}

\item{legend}{Logical, if TRUE add a legend}

\item{query}{Show a low/high profile for certain markers in the legend.
See also \code{\link{QueryStarPlot}}}

\item{range}{If "all" (default), range is computed on all markers passed,
if "one", range is computed on every marker separately. The
height of the pie pieces will be computed relative to this 
range.}

\item{main}{Title of the plot}
}
\value{
Nothing is returned. A plot is drawn in which each node is 
represented by a star chart indicating the median fluorescence intensities.
Resets the layout back to 1 plot at the end.
}
\description{
Plot FlowSOM grid or tree, where each node is represented by 
a star chart indicating median marker values
}
\examples{
   # Read from file, build self-organizing map and minimal spanning tree
   fileName <- system.file("extdata", "68983.fcs", package="FlowSOM")
   flowSOM.res <- ReadInput(fileName, compensate=TRUE,transform=TRUE,
                            scale=TRUE)
   flowSOM.res <- BuildSOM(flowSOM.res,colsToUse=c(9,12,14:18))
   flowSOM.res <- BuildMST(flowSOM.res)
   
   # Plot stars indicating the MFI of the cells present in the nodes
   PlotStars(flowSOM.res)

}
\seealso{
\code{\link{PlotPies}},\code{\link{PlotMarker}},
\code{\link{PlotCenters}}, \code{\link{BuildMST}}
}
