% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_buildSOM.R
\name{SaveClustersToFCS}
\alias{SaveClustersToFCS}
\title{Write FlowSOM clustering results to the original FCS files}
\usage{
SaveClustersToFCS(
  fsom,
  original_files,
  pp_files = original_files,
  selection_files = NULL,
  silent = FALSE
)
}
\arguments{
\item{fsom}{FlowSOM object as generated by BuildSOM}

\item{original_files}{FCS files that should be extended}

\item{pp_files}{FCS files that correspond to the input of FlowSOM}

\item{selection_files}{Files indicating which cells of the original files
correspond to the input files}

\item{silent}{If FALSE (default), print some extra output}
}
\value{
Saves the extended fcs file as [originalName]_FlowSOM.fcs
}
\description{
Write FlowSOM clustering results to the original FCS files
}
